Color Options and Grid Bug Fix
There are now more options for colors: when you click on the left or
right color, a color chooser pops up that gives more color options
including transparency. There was a bug where the grid was only drawn
for the primary image in tiled mode, but it has been fixed.